refactor: 역할 분리를 위한 리펙토링
